% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{alt,TasselHDF5-method}
\alias{alt,TasselHDF5-method}
\title{Accessor for alterate alleles from a TasselHDF5 object}
\usage{
\S4method{alt}{TasselHDF5}(x, as_char = TRUE)
}
\arguments{
\item{x}{a TasselHDF5 object}

\item{as_char}{logical indicating whether to return as \code{CharacterList}
or as Tassel encodes them, as an \code{IntegerList}}
}
\description{
Accessor for alterate alleles from a TasselHDF5 object
}

