% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{geno,TasselHDF5-method}
\alias{geno,TasselHDF5-method}
\title{Accessor for genotype information from a TasselHDF5 object}
\usage{
\S4method{geno}{TasselHDF5}(x)
}
\arguments{
\item{x}{a TasselHDF5 object}
}
\description{
Accessor for genotype information from a TasselHDF5 object
}

