% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hdf5load.R
\name{initTasselHDF5}
\alias{initTasselHDF5}
\title{Initialize a TasselHDF5 object, connected to an HDF5 file}
\usage{
initTasselHDF5(file, version = "5")
}
\arguments{
\item{file}{a path to an HDF5 file}
}
\description{
Initialize a TasselHDF5 object, connected to an HDF5 file
}

