% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hdf5load.R
\docType{methods}
\name{loadBiallelicGenotypes,TasselHDF5-method}
\alias{loadBiallelicGenotypes,TasselHDF5-method}
\title{Load and decode biallelic genotypes from HDF5 file}
\usage{
\S4method{loadBiallelicGenotypes}{TasselHDF5}(x, verbose = TRUE)
}
\arguments{
\item{x}{a TasselHDF5 object}

\item{verbose}{a logical describing whether to be verbose during loading}
}
\description{
Load and decode biallelic genotypes from HDF5 file
}

