% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{loadBiallelicGenotypes}
\alias{loadBiallelicGenotypes}
\title{Load biallelic genotype information}
\usage{
loadBiallelicGenotypes(x, verbose = TRUE)
}
\arguments{
\item{x}{a TasselHDF5 object}

\item{verbose}{logical indicating whether to print status mesages during loading}
}
\description{
Load biallelic genotype information
}

