% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{ref,TasselHDF5-method}
\alias{ref,TasselHDF5-method}
\title{Accessor for reference alleles from a TasselHDF5 object}
\usage{
\S4method{ref}{TasselHDF5}(x, as_char = TRUE)
}
\arguments{
\item{x}{a TasselHDF5 object}

\item{as_char}{logical indicating whether to return as character
or as Tassel encodes them, as integer}
}
\description{
Accessor for reference alleles from a TasselHDF5 object
}

