% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{ref}
\alias{ref}
\title{Accessor for reference alleles}
\usage{
ref(x, as_char = TRUE)
}
\arguments{
\item{x}{a TasselHDF5 object}

\item{as_char}{logical indicating whether to return as character
or as Tassel encodes them, as integer}
}
\description{
Accessor for reference alleles
}

