% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tasselhdf5-methods.R
\docType{methods}
\name{samples,TasselHDF5-method}
\alias{samples,TasselHDF5-method}
\title{Accessor for samples from a TasselHDF5 object}
\usage{
\S4method{samples}{TasselHDF5}(object)
}
\arguments{
\item{object}{a TasselHDF5 object}
}
\description{
Accessor for samples from a TasselHDF5 object
}

