% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tasselhdf5-methods.R
\docType{methods}
\name{granges,TasselHDF5-method}
\alias{granges,TasselHDF5-method}
\title{Accessor for genomic ranges from a TasselHDF5 object}
\usage{
\S4method{granges}{TasselHDF5}(x, use.mcols = FALSE, ...)
}
\arguments{
\item{x}{a TasselHDF5 object}
}
\description{
Accessor for genomic ranges from a TasselHDF5 object
}

