% Generated by roxygen2 (4.0.1): do not edit by hand
\name{initTasselHDF5}
\alias{initTasselHDF5}
\title{Initialize a TasselHDF5 object, connected to an HDF5 file}
\usage{
initTasselHDF5(file)
}
\arguments{
\item{file}{a path to an HDF5 file}
}
\description{
Initialize a TasselHDF5 object, connected to an HDF5 file
}

