% Generated by roxygen2 (4.0.1): do not edit by hand
\name{loadBiallelicGenotypes}
\alias{loadBiallelicGenotypes}
\title{Load and decode biallelic genotypes from HDF5 file}
\usage{
loadBiallelicGenotypes(x, verbose = TRUE)
}
\arguments{
\item{x}{a TasselHDF5 object}

\item{verbose}{a logical describing whether to be verbose during loading}
}
\description{
Load and decode biallelic genotypes from HDF5 file
}

