% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{show,TasselHDF5-method}
\alias{show,TasselHDF5-method}
\title{Pretty-print a TasselHDF5 object}
\usage{
\S4method{show}{TasselHDF5}(object)
}
\arguments{
\item{object}{a TasselHDF5 object}
}
\description{
Pretty-print a TasselHDF5 object
}

